refactor: remove the fast server
