+++ .dockerignore
@@ -3,10 +3,7 @@
 config/postal.yml
 config/smtp.cert
 config/smtp.key
-config/lets_encrypt.pem
 config/signing.key
-config/fast_server.cert
-config/fast_server.key
 public/assets
 vendor/bundle
 Procfile.local
