+++ .gitignore
@@ -19,10 +19,7 @@
 config/postal.yml
 config/smtp.cert
 config/smtp.key
-config/lets_encrypt.pem
 config/signing.key
-config/fast_server.cert
-config/fast_server.key
 config/postal/**/*
 
 spec/config/postal.local.yml
