+++ Procfile
@@ -1,5 +1,4 @@
 web: bundle exec puma -C config/puma.rb
-fast: bundle exec rake postal:fast_server
 worker: bundle exec ruby script/worker.rb
 cron: bundle exec rake postal:cron
 smtp: bundle exec rake postal:smtp_server
